Add newsApiTransform class for transforming news data
This commit adds a new file, newsApiTransform.js, which contains a class called newsApiTransform. The class has a static method called transform that takes in a news object and returns a transformed version of the data. The transformed data includes the id, heading, news content, and image of the news article. This class will be used to transform news data fetched from an API before storing it in the database.