refactor: Add removeImage function to helper.js
The code changes in `helper.js` add a new function called `removeImage` that is responsible for removing an image file from the file system. This function takes an `imageName` parameter and constructs the path to the image file using the `process.cwd()` and `"/public/images/"` variables. If the file exists at the specified path, it is deleted using the `fs.unlinkSync` method. This addition enhances the functionality of the application by providing a convenient way to remove images from the file system.