Add uploadImage function to helper.js
The code changes in `helper.js` add a new function called `uploadImage` that is responsible for uploading an image file to the specified directory. This function takes an `image` parameter and generates a random image name using the `generateRandomNum` function. The image is then moved to the upload path using the `mv` method. This addition enhances the functionality of the application by providing a convenient way to upload images.