Add Redis caching for news route
This commit adds Redis caching to the `/news` route in `api.js`. By using the `redisCache` middleware, the news data will be cached, improving the performance of the application. This enhancement leverages the recently added `redis.config.js` file, which contains the necessary configuration for using Redis as a caching mechanism. The caching mechanism will store the news data for a specified expiration time, allowing subsequent requests to be served from the cache instead of querying the database. This optimization will reduce the load on the database and improve the overall response time of the `/news` route.